This PR introduces a custom `Scroller` class that enhances the scrolling speed of the `RecyclerView`. The new implementation uses a quadratic algorithm to increase the scrolling speed based on the distance already scrolled, resulting in a more responsive and faster scrolling experience. The quadratic formula parameters can always be adjusted to fine-tune the scrolling speed.

Previously:
~30 seconds scrolling down -> ~19 seconds to scroll back up

Now:
~30 seconds scrolling down -> ~9 seconds to scroll back up

Closes #195